Fix swiping up bouncer on GH
The bouncer gesture is being interrupted by a call from
SharedNotificationContainerViewModel to update the max number of
notifications, which has the downstream effect of hiding the bouncer.
SharedNotificationContainerViewModel won't update the max number of
notifications if the user is interacting with the shade however, but we
don't currently set that state in the GH bouncer touch handling. This
change updates the user gesture state in ShadeRepository so that all
downstream consumers know the user is interacting with the shade when
expanding the bouncer on GH

Fixes: 391411581
Test: atest SystemUiRoboTests:BouncerSwipeTouchHandlerTest
Flag: EXEMPT bugfix
Change-Id: Id66dee4d24bb7e8754d6f887e84b342e23121d93

Former-commit-id: 4cc106f8908d186d05863d7134f3ef481520a774