[A11y] Fix bug where magnification jumps between zoom levels
Fullscreen magnification would get into a state where it continuously
jumped between zoom levels when a pinch zoom or keyboard shortcut
zoom came in at the same time as the window magnification settings
panel was opened. This can be fixed by passing in 'fromUser',
which indicates that the user did not initiate the change and it
can be ignored.

Tested manually with TalkBack to ensure that TalkBack changes to
the progress bar (using the swipe up/down gestures) still cause
magnification scale changes. TalkBack interacting with the +/-
buttons also still cause scale changes.

Bug: 392946442
Test: atest SeekBarWithIconButtonsViewTest
Test: atest WindowMagnificationSettingsTest
Test: Manual with TalkBack
Flag: EXEMPT bugfix
Change-Id: I0d0b925bb00562fa7bec187e59702c28429b4e07

Former-commit-id: 1f8d924a37bfcb302ca55d20b2a2992594b23eae