Fix wrong compat override screen sizes
Override is applied when the insets decouple is enabled and the app is a
legacy app. When calculating the screenWidthDp and screenHeightDp, the
display size excluding status bars, navigation bars, and display cutouts
should be used.

The current implementation is calculating the screenWidthDp and
screenHeightDp by excluding the navigation bars and display cutouts from
the display size. This is not consistent with the legacy behavior of the
system.

When the device has a status bar larger than a display cutout, or the
device doesn't have a display cutout, some app may experience layout
issue because of the inconsistency.

Bug: 391063639
Test: WindowProcessControllerTests
Flag: EXEMPT bugfix
Change-Id: I2e3ac42ac341e8eff7a01dd4985451b7e37a4842

Former-commit-id: bf8b16bc6cf96fe58049a3b525c773d97d9b7ace