Update Kosmos Fixtures for CoroutineContext
These Fixtures were incorrectly pulling the context from the TestScope,
which could cause unexpected behavior due to including the Job from the
TestScope.

In reality, these Contexts are meant to be used to launch code on the
appropriate thread (bg vs main), and so have been updated to just point
to the testDispatcher, which accomplishes this.

Flag: EXEMPT tests only
Test: atest SystemUITests
Change-Id: If0aa708f8ce677f5745b5967f83be28de6679246

Former-commit-id: 3ca3c472395b7defea9e52dba51c8d256176fda7