Fix NullPointerException in SettingsHelperRestoreTest.
Some cases in SettingsHelperRestoreTest were crashing due to a NPE with the logger. This didn't happen in the actual code or in SettingsHelperTest because the logger was set when the flag was on, but I missed to do this in SettingsHelperRestoreTest. This CL sets the logger in the test, and checks for null in the code to be extra safe.

Change-Id: I8d5c0fdd9901e889702570b8433853e4f1dbd692
Bug: 379861078
Bug: 394349769
Flag: com.android.server.backup.enable_metrics_settings_backup_agents
Tested: atest SettingsHelperTest
Tested: atest SettingsRestoreHelperTest

Former-commit-id: 40b199020fdfa1111a73134babd52eac1f5c50ed