Add the supervision profile type.
This new user type is based on the parentless profile used by the
communal profile. This profile can be started in the background by any
user and can show UI over any user. It is meant to run very infrequently
only for actions involved in setting the supervision credentials.

This CL just allows for creating the supervision profile, but does not
yet introduce changes required for the keyboard to work correctly for
activities running in the context of this user. Future work will also
limit the number of applications installed into the profile.

Test: atest com.android.server.pm.UserManagerTest
Bug: 389712089
Flag: android.multiuser.allow_supervising_profile
Change-Id: Ic3fe1c2b11b24dd709721a858e88adf2cdb1e640

Former-commit-id: 44216415f96664a8a7083d39ddbb4d3096ba7080