[sb] requestLayout if the number of levels changes
With new_status_bar_icons enabled, the LevelList drawable that holds the
mobile icons contains two main widths of icon: the 4-bar and 5-bar
icons.

Along with the change to SignalDrawable that no longer sets a square
intrinsicWidth/Height, this change is required to ensure that the layout
of the icons in the StatusIconContainer get the most-up-to-date width
from the icon.

Test: restart sysui with a device that uses 5-bar SIM, and with Wi-FI
on. Without the fix, the wi-fi icon will elide into a dot
Bug: 391606042
Flag: com.android.settingslib.flags.new_status_bar_icons

Change-Id: I082cae5749048d53682206e8a6b4d723f60b54d3

Former-commit-id: 6015cc0d688f7054a0940395631b2bed94f1828b