Re enable IME tests for show/hide flags
Previously we disabled IME tests that relied on behaviour specific to
flags like SHOW_IMPLICIT, SHOW_FORCED, HIDE_IMPLICIT_ONLY or
HIDE_NOT_ALWAYS, as the IME insets controller refactor made it so we
lose this flag information.

This re-enables the tests, verifying that having the flags passed has
the same behaviour as having no flags set, when the refactor is enabled.

Flag: EXEMPT testfix
Bug: 394328311
Test: atest InputMethodServiceTest
Change-Id: I29a01e1bb00ae841270a24465cbcbb1a031c0f3b

Former-commit-id: 69508eee4c9e4051309e323e697711b8c9785b7d