Allow freeform openInstance to handle non-freeform running tasks.
openInstance has been functioning on the assumption that when the calling task is freeform and the requested task is not freeform, the requested task must be in the background. This is not necessarily the case and leads to incorrect bounds on launch. This CL routes the logic through moveTaskToDesk to correct this assumption.

Bug: 395975053
Test: Manual
Flag: com.android.window.flags.enable_desktop_windowing_multi_instance_features
Change-Id: I03ebc4e0d446db7401bfb1d06318fa789b8a5d1d

Former-commit-id: 01ca704c9ab037b27608b526b816d1c24d917712