Relax integer limits in argument handling of native Zygote loop
Previously, the picky version of atoi() would reject anything 7 digits
or more. We have many cases of 7-digit UID these days, so we need a
higher limit to make good use of native Zygote loop.

Relax this to allow parsing up to the integer type limit. This function
is only used for --setuid and --setgid parsing, and the corresponding
Java code also uses Integer.parseInt(). So we should end up having the
same limits as that (except that ours refuses to parse negative values).

Bug: 395628427
Test: manual; check absence of "forkRepeatedly terminated due to non-simple command" with Chrome
Change-Id: I621080d618a224460603d2318dd9984120c37cb4

Former-commit-id: 73075880787bd355de7ae7e1cc82967b922624d2