[ROSP] Make Media Chip functional without flexiglass
This CL adds a new mediaControlChipModelLegacy flow in
MediaControlChipInteractor which directly listens to the
MediaDataManager to provide a MediaControlChipModel. This allows the
chip to function without flexiglass being enabled.

Also, updates the MediaControlChipInteractor to handle both flexiglass
on/off scenarios and creates a new isEnabled field which is used to
ensure this feature is only enabled on SystemUITitan. This is done by
setting the isEnabled field in a MediaControlChipStartable.

Bug: b/394887130
Flag: com.android.systemui.status_bar_popup_chips

Test: parameterized
tests for MediaControlChipInteractor
Test: Make sure media control chip
is functional with flexiglass on/off.

Change-Id: I21db9377eed72b1f485df3d0d49adb71149536b1
Former-commit-id: fef3243e5da8f558a1ebc4a8c7276aa2e6a04857