Fix time overlap issue in Sign-in Options screen of Credential Manager
There is an issue of time overlapping with other content as the user
scrolls in Sign-in options screen of Address book app in Credential
Manager.

This change fixes the issue by changing the columnState of
MultiCredentialsFlattenScreen from being set to be laid out directly
below the time text to being passed as a parameter through the
scrollable just like the other screens.

Bug: 377251580
Test: Open Address Book -> Login -> Sign-in options. Scroll through
the screen and check that the time text at the top doesn't overlap
with other content in the screen. Screenrecording of the fix attached
in the bug.

Change-Id: I9bf964a6783dfb088f9f74ee063de7d75386e8f8

Former-commit-id: 80d4783cc58f84bec854b93100ed7c2bab60128c