Correctly initialize virtual display power state
Effectively removing the usages of FLAG_NEVER_BLANK. This flag was
added in 2013 (ag/316512) but all the display blanking logic from
back then is gone. The only thing the flag does today is prevent
calls to requestDisplayStateLocked for virtual displays.

Removing this allows to set the initial state of a display to UNKNOWN
because requestDisplayStateLocked is always called now.

So calling requestDisplayStateLocked will trigger the
VirtualDisplay.Callback on creation (which is what should happen) and
will trigger onDisplayChanged callback for DisplayListeners.

Fix: 371125136
Test: presubmit, CTS
Flag: android.companion.virtualdevice.flags.correct_virtual_display_power_state
Change-Id: I82f6bf6b4a8e56f8b59ba601b406982642f5d503

Former-commit-id: 519bd4e7f56290a069de08718760717f68fe3c99