[sb] calculate overall chip visibilty in view model
When StatusBarRootModernization is on, we need to account for the
overall status bar visibility (shouldHomeStatusBarBeVisible), as well as
the possibility that a HUN is showing, when it comes to the chip
visibility.

This CL exposes a new property (canShowOngoingActivityChips) that covers
this.

It also handles an edge case where the binder might try to set something
as visible via `show()`, but the view is already visible with alpha=0.
In this case, we still want to run the show function since the alpha
needs to be updated.

Test: HomeStatusBarViewModelImplTest
Test: have ongoing call notification and go to lockscreen
Fixes: 395632689
Flag: com.android.systemui.status_bar_root_modernization
Change-Id: I91bb7e552b20895ad2e9d3848072a7779e9ff5e5

Former-commit-id: 3df026673090885f6126fad0c15b34dde206627b