Send KEYCODE_POWER instead of KEYCODE_TV_POWER.
From b/291854577, KEYCODE_TV_POWER has some problem as interoperability issues where the TV doesn't act on the HDMI-CEC messages as expected. Also, if the set-top box's and TV's power state are not in sync (one is ON and the other is OFF), the power control could be wrong.
From https://developer.android.com/reference/android/view/KeyEvent#KEYCODE_TV_POWER, it seems this is only for like set-top box to use HDMI-CEC to control TV power or not. But both TV and STB support KEYCODE_POWER.

Bug: 390553665
Change-Id: If664d7a6dd680579c7e6d69e1fcd900edb3530f0
Flag: EXEMPT bugfix

Former-commit-id: c95e6e175a7917e094debb9b31c6bd371a8b65ef