Drag-to-desktop transition: attach visual indicator to transition root
Attach the visual indicator to the drag-to-desktop transition root, so
the visual indicator is removed at the end of the transition regardless
of whether the indicator was released properly (we might not release the
indicator properly if the SysUI process crashes).

Note: attaching the visual indicator to the transition root only works
for the drag-to-desktop case. In the case where we drag a window inside
Desktop Mode we are not running a transition so we don't have a
transition root.

Test: used winscope to ensure indicator lives under transition root
Bug: 392826275
Flag: com.android.window.flags.enable_visual_indicator_in_transition_bugfix

Change-Id: I5d608fd11aad05addf14f06be1a2dd8d30bc8b9b

Former-commit-id: 38a5db4f6d0040bfcbdc90334c08ad59ab7622ce