Let window overwrite insets again if it has a (fake) control
When a heads up notification was active, and we started to hide insets types, this would invoke the onApplyWindowInsets callback twice. This happened due to the fact that we hold a fake control, which is replaced by the real control, after the notification is gone.
Previously, overwriting the insets was only possible, if the control had a leash, which is reverted by this CL.

Fix: 395816479
Flag: EXEMPT bugfix
Test: atest android.server.wm.insets.WindowInsetsControllerTests#testDispatchApplyWindowInsetsCount_systemBars
Change-Id: I2523607e409a2d86b7d7f58e46e0fe6e29a47ad9

Former-commit-id: b1c0a9dee8609e4341c7d2f1dd140414f7832acc