Fix display events emitted when override info changed
After the split of display changed events into basic changes and refresh
rate changes, we have continued to emit both these events when
non-overridden display info is checked. This is incorrect as it is
possible for displays's basic information to change but refresh rate
still remains the same.

Bug: 397020451
Flag: EXEMPT trivial change
Test: Manually run abtd to validate the metrics
Change-Id: I14c031acb3fac4ac38d75c1cccfd2884bfbef005

Former-commit-id: a7e006da5cc5ee819d40eea71edaf3758c272e4f