Fix swipe-to-bouncer on GH with face auth
When swiping up the bouncer on glanceable hub, if the user has a large
number of notifications, the notification limit is taking effect
mid-swipe and causing the bouncer to hide itself. This is because the
bouncer/alternate bouncer states are considered "lockscreen" states,
even if the bouncer is shown over the glanceable hub instead of over the
lockscreen.

This change does not consider the bouncer states as "lockscreen" if they
are being shown over the glanceable hub.

Fixes: 396689194
Test: atest SharedNotificationContainerViewModelTest
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I8cfa5ff3074d7d34f06b7db5a1051e367e290889

Former-commit-id: fecfb9756dcebd824e74637a75bb623cc30846cc