Handle UserSwitchingDialog animation timeouts on a non-ui thread.
Animations of the dialog already use the ui thread. And when something
is wrong there -like being very busy at the time- handling the
mHandler.postDelayed calls on the same thread does not work.
This CL moves timeout posts to be handled by ActivityManager thread,
instead of the ui thread.

Bug: 393514824
Test: atest UserControllerTest
Flag: EXEMPT bugfix
Change-Id: I003db7c8ce8b310a9386de74055a55db9775f5fa

Former-commit-id: e1d31fe34acd408ef816ec12178c60cb818aa960