[Notif redesign] Update notification info design
Updating the layout of notification guts to match the redesign specs.

Note that notification_2025_info.xml is a fork of notification_info.xml,
and notification_2025_guts_priority_button_bg.xml  is a fork of
notification_guts_priority_button_bg.xml.

There's a lot of duplication in this layout that is less than ideal, but
I left it alone since we plan to rewrite this in compose very soon
anyway. I will also remove the "Automatic" button in a follow-up because
we're not actually using it.

Bug: 394822197
Test: tested manually
Flag: com.android.systemui.notifications_redesign_guts
Change-Id: Ibb7c54c6cbede0ef5426315cd9f9753ee55e1806

Former-commit-id: 98a6e637d7faa38091e432c5604b4c17a752defd