[35/N] Desks: Add DesksOrganizer#unminimizeTask
Unminimizes a task from a desk if it was minimized. Internally, it means
the task is reparented from the minimization root to the desk root.

To a DesksOrganizer, unminimization does not imply desk activation, so
activating an inactive desk by unminimizing one of its tasks will still
need to be explicitly done by DesktopTasksController in a separate CL.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 393978427
Test: atest WMShellUnitTests
Change-Id: Ic6d9c4b2c1731c4edcdd184dabaf34d17d155253

Former-commit-id: 4839d2958e624684a317345ab3d3e0f0cc435a90