Pass volume with correct user id to StorageSessionController
When remounting public volumes, we pass the volume after setting its
mount user id to the new user, it's going to be mounted to. This is
incorrect, as it doesn't really allow StorageSessionController
clean up the old session from the old user, but tries to clean it up
from the new user which doesn't do thing at the end.

Luckily, we rely on vold to do the real logic of unmounting fuse paths,
which in turn takes care of existing the corresponding fuse daemon.
But the incorrect clean up could cause some unnecessary delays
when further reset operation happens, as we try to kill stale sessions.

Bug: 327173406
Bug: 369519866
Test: atest StorageManagerService & manual
Flag: EXEMPT bug fix
Change-Id: Ie4851f316ea384433f97da84e47cbcbf4c367a7a

Former-commit-id: 8c44db9e6a31365d2a15ab3993f132fe7d395069