Remove from/to attributes from libcore.util.Nullable/NonNull
The attributes are not used, either in the code, or in Metalava and
their presence prevents mapping them to other, more common nullability
annotations like those in `android.annotation` or `androidx.annotation`
as Metalava needs to do when generating the stubs.

Matches the changes made in https://r.android.com/3508044.

Bug: 311619997
Test: m checkbuild
Change-Id: I4af4aa79f973dab41cf22c968fa83fe90d095610

Former-commit-id: d29b9cdc8e4061321fa602baac1382c408a1c810