Propagate Notification and QS SysUI flags per display
NPVC now propagates SysUIstate changes through SysUIStateDisplaysInteractor. This guarantees those flags are only on for a specific display, and off for all the others.

The case when SceneContainer flag is on has been handled already by one of the cls this is based on.

Bug: 362719719
Bug: 398011576
Test: SysUIStateDisplaysInteractorTest, NotificationPanelViewControllerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I0d1044718fae3933cc790d3bc4ca01dbb19ecad0

Former-commit-id: 7d55651b5614ffc98c9c60d775f3f11f52ed12c4