Check shade expansion fraction to allow animation
In compose, it seems that values doesn't get updated before the time of
the transition of UMO from QS to QQS. Therefore, the UMO is still marked
as if it's visible and animation is executed, which leads to the flicker
effect. Checking the shade expanded fraction and decide whether
animation is needed fixes the issue.

Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Fixes: 389576164
Test: Checked UI.
Change-Id: I741f5118ab9d1f66cd10dfe236a374184b3908b6

Former-commit-id: 894de67e572e9750b206d476961848f313f49206