Channel SysUI state updates through SysUIStateDispatcher
After this change, any listener registered to any SysUIState instance will receive updates for all display IDs.

The logic to handle callbacks is removed from SysUIState and moved to the dispatcher (that is also thread safe now).

This is needed as various parts of the code will change the state for different displays, and we want each listener to get events for all displays.

In a follow up it's possible to entirely remove callback registration logic from SysUIState and have it only from a repository, but it was a bigger refactor, and I prefer to keep this cl small.

Bug: 362719719
Bug: 398011576
Test: SysUIStateDispatcherTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I465b207a3d2ab28d7cd66f09e98c013160532d1d

Former-commit-id: c7d2e77a57be4d7ea06931c7c0b6d503683b5559