Add scene.alwaysCompose flag to STL scenes
This CL adds an alwaysCompose flag that can be added to any STL scene so
that this scene is always composed. Instead of being removed from
composition when they leave the current STL state/transition, these
scenes are made invisible by not placing them. That way they are not
drawn and don't participate in gesture handling, but they are still
composed and measured. This can help reduce the time it takes to
transition to a scene, and will be used for the Compose QS in
particular. See b/389985793#comment109 for details.

Important: This is not meant to be used outside of Compose QS, and
should not be used by any Flexiglass scene.

Bug: 389985793
Bug: 399825091
Test: atest SceneTransitionLayoutTest
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Change-Id: I17462c64fc2ae5e70963ec4b1129ddb03930d287

Former-commit-id: d9bcdd72f652dac49f1f398a2130f1fc1915a791