[SB][Notif] Make sure re-tapping notif chip hides HUN immediately.
I originally implemented this behavior in ag/31139202 and it seemed like
it worked because the minimum timeout for the HUN was 1 second. But
later in ag/31744914 we updated the minimum timeout for chip HUNs to be
3 seconds, which means re-tapping the chip no longer worked if you did
it within the first 3 seconds.

This CL ensures we remove the HUN immediately if the user re-tapped the
chip since it's a clear user action.

Fixes: 399882113
Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips
Test: Tap chip then immediately re-tap chip -> verify HUN immediately
disappears (see video in bug)
Test: atest HeadsUpCoordinatorTest

Change-Id: I62e71c425df8d05cf2771ea2731d9ad1661d69f9

Former-commit-id: 02e6e7e33afb846db0624d5399dd5f856a7c27e4