Do not load DecorView info if a task is closing
When an app is removed with split screen activated, trying to find the
task and keeping the split screen is not feasible. In that case, we
should not relayout the corresponding DecorView as it has no effect and
will crash. With this change, removing an app in split screen will no
longer cause crashes.

Bug: 394502142
Test: Remove app while the app is in split screen, observes no crash
Flag: com.android.window.flags.skip_decor_view_relayout_when_closing_bugfix
Change-Id: Iceeec260317b8af97fa3538e53e0f896241855e7

Former-commit-id: b610010fd5ba143e4dad687eed357210925b345d