[Notif redesign] Update large icon spacing
In the new design, there should be 16dp of space between the content and
the right icon, and 8dp between the icon and the expander pill. That
means that we need to update the icon margins and dynamically adjust the
padding of the expand button.

This also fixes the long-standing issue of the icon moving around when a
notification is expanded/collapsed.

Note that notification_2025_right_icon is a fork of
notification_template_right_icon.

Bug: 378660052
Test: screenshot tests
Flag: android.app.notifications_redesign_templates
Change-Id: I716ece1a4a42406f4bfd1dd3ce0bf55e8d084db4

Former-commit-id: ad51fe08985debee6668f2c41ec7fcc322117dc8