Add "Optimize View" option to Handle Menu
This is the new app-compat UI element to give users the option
to restart apps when they move between displays and get some
app-specific UI issues. (See the mock attached in the bug for
more details.)

The visibility of the menu is controlled by
AppCompatTaskInfo.isRestartMenuEnabledForDisplayMove, which
becomes true when an app moves between displays, and reset to
false when an app actually gets restarted.

When the button is pressed, it invokes
CompatUIHandler#sendCompatUIRequest() and the restart dialog gets
shown.

Flag: com.android.window.flags.enable_restart_menu_for_connected_displays
Bug: 400575337
Test: DesktopModeWindowDecorationTests
Change-Id: I398b01fe044ccdc960603a0b9391eb90e16575b3

Former-commit-id: 187d2a313c6f07e67778260103dced0189f3eb12