Ensure SystemFeaturesMetadata is initialized in the boot image
While the overhead of recomputing the feature lookup cache is relatively
cheap, we should avoid it by computing it just once with the boot image.

Note that this manual addition should only be necessary as we ramp the
feature, after that we can likely remove it and rely on the usual boot
image profiling pipeline to ensure one-time init.

Bug: 400597432
Test: m + check boot-framework.art/oat
Flag: android.content.pm.cache_sdk_system_features
Change-Id: I1f4e881c94a3a4c1b7c1bedda0d3968ce47a71d2

Former-commit-id: 23b4f532b1b5126ea5c41f119b53a9da792197a3