Check wiu permissions on any deviceId to start a fgs
Some of the wiu permissions required to start a foreground service can
be device aware (RECORD_AUDIO and CAMERA) and they can be different
depending on the resources (virtual audio or camera devices) used.
This relaxes the check to allow the fgs to start if it has the permission
on any of the active deviceIds (virtual devices).

CTS test added in ag/32131429, @Ignore for the immediate release (though
it passes and will be reenabled by b/402139346).

Bug: 381386115
Bug: 400753852
Test: manual VDM demo sample app with Pixel Recorder app
Test: atest android.virtualdevice.cts.audio.VirtualAudioPermissionTest
Test: atest ActivityManagerForegroundServiceTypeTest
Flag: EXEMPT, bugfix
Change-Id: I91b01ba2dddda83801a2a9b586cd03bbdc780754

Former-commit-id: 49156faf47fd4c72529420e78b9eb3f7b709e796