Ensure that DeferredDisplayUpdater logs are printed to logcat
Use WM_DEBUG_WINDOW_TRANSITIONS_MIN to ensure that the logs
in DeferredDisplayUpdater are also printed to logcat, as
we had several bugreports with missing protologs.

Also adding width and height of the DisplayInfo object which
might help to investigate a problem when the device is stuck with
incorrectly applied DisplayInfo (for a wrong screen).

Test: manually checked that logs printed to logcat
Flag: EXEMPT bugfix
Bug: 362548632
Change-Id: Ic7b5e92eb9bd429e7bf04cd3a99ecff172c282f3

Former-commit-id: b0b7989acf13a48a7cbe4167ee6f46fc333a5c1a