Disable blur forcing shade window to be visible
ag/10672056 added a check for a non-zero blur radius in the logic
that determines whether the window should be visible. We believe
that the check is not required for the new implementation of blur
and that a race condition that only occurs on emulated devices
is causing tests to flake due to non-zero blur being set on an
invisible shade, which makes it become visible and steal input
events.

Bug: 394977231
Flag: com.android.systemui.disable_shade_visible_with_blur
Test: manually checked shade CUJs
Change-Id: Ie433a4378fea6ccc81f942a50b1eb71ee14cf763

Former-commit-id: 6a3ad1a18c873b1893d88a98f9a3ba9ff5cdc40d