[kairos] Have icon adapter track latest interactor
Presently, various view-models depend on the interactor returned from
MobileIconsInteractor#getMobileConnectionInteractorForSubId remaining
"alive" indefinitely, even if the connection disappears and returns.
This is due to a "desync" between the set of active connections and the
set of subscriptionIds.

When the ui layer is migrated to Kairos this will no longer be an issue,
because the new Kairos versions expose IconInteractors directly as
opposed to the indirection of the subscription id. For now though, the adapter must respect the behavior of the original API.

Flag: com.android.systemui.status_bar_mobile_icon_kairos
Bug: 383172066
Test: atest
Change-Id: Ic5a82b6df6723fa423dbc6b4deb7982ad5727068

Former-commit-id: 470592229ec585bceb16584dd3999bce12b6e70a