[pm] block install/uninstall shell commands before boot_completed
We've observed in automated testing environments that occasionally a shell command to install/uninstall packages can be sent to PackageManagerService right after it is started during boot, but before the rest of the system (art service, broadcast controller, etc) is ready. The system crashes when PackageManagerService tries to process such install/uninstall requests and invokes other services that are not ready. This CL prevents that by checking if the system has finished booting before processing the shell commands.

BUG: 370100381
FIXES: 370100381
BUG: 369993707
FIXES: 369993707
Test: manually install/uninstall via shell commands during boot and observe "Error: device is still booting."
FLAG: EXEMPT bug fix
Change-Id: I74759ceaacf59a99507256572b9ef1680a942e8d

Former-commit-id: e38fa24e5738513d721ec2d9fd2dd00f32e327c1