Clear topTransparentFullscreenTask from repository if another task opens
The DesktopTransistionObserver does not receive the close or to back
transition for the top transparent fullscreen task causing it to not be
cleared correctly. This leads to freeform windows opening outside of
desktop as we think desktop stop is still showing.

Instead, if the top transparent fullscreen task has not been cleared as
another task opens, or is brought to front, clear the top transparent
fullscreen task as it is no longer needed.

Flag: com.android.window.flags.include_top_transparent_fullscreen_task_in_desktop_heuristic
Fixes: 398167855

Test: atest WMShellUnitTests:DesktopTasksTransitionObserverTest
Change-Id: I737d24c1f143843270f554d4828b7f6a5fce063c

Former-commit-id: 666c6b12a50d731dd2f7cc73c1a8b685dc67e8fb