Do not create dim layer until it gets a valid dimming request
If a window requests a dim of 0 and no blur, the dim surface gets created (because
technically someone is requesting to dim) and then immediately
destroyed (because alpha=0, blur=0 means that this layer should be destroyed as
it's useless). We can postpone creating a new surface until the window
requests a non 0 value.

Bug: 401449608
Test: DimmerTests
Flag: EXEMPT minor change
Change-Id: I11c97ae6401e3151df4941a5ec6ad42bb94357ce

Former-commit-id: 89341c12f60b1253327748c2ef8fbc43e927eb1e