Introduce TransitionManager and functioning return animations to call chips.
This still has some gaps that will be addressed in follow-up CLs, but I
tried to split up the code as much as I could to avoid making it too hard
to review at once.

Additionally, some of the creation code for the factory and transition
manager will probably have to move once we add this functionality to RONs
as well. However I don't have a clear path forward there yet, so I would
rather get the code in for the call chip first and revise later.

Bug: 202516970
Flag: com.android.systemui.status_bar_chips_return_animations
Test: atest CallChipViewModelTest
Change-Id: If945338276647bbfd83bf8949af85cefcd41cec1

Former-commit-id: 63bada69ea3be3a0c2388fa276c8c82de8079aef