[Media] Content revealing pager.
Adds support for a horizontal pager that doesn't dismiss. Instead, it
reveals a box with a spinning/fading settings icon that can be clicked
to open media settings.

The CL adds the new ContentRevealingHorizontalPager, cleans up the
previously-existing swipe-to-dismiss pager in
DismissibleHorizontalPager, and ties it all in Media.kt.

The CL also includes minimal changes to NestedDraggable and
OffsetOverscrollEffect that are necessary to support the new
swipe-to-reveal pager.

Bug: 397989775
Test: manually verified both swipe-to-dismiss and to-reveal in the
Compose gallery app
Flag: EXEMPT code isn't yet used in production

Change-Id: I216c5df586c959c1d1adddd78ddfa5c47632ede0

Former-commit-id: 94ba5f0127ee4ed616ec87d12e3c34fe22bdde8d