Include bar outline for bubble drag indicators
When dragging a task to bubble we show the expanded view hint as the
drop target.
We also need to include an oval for the bubble bar to better emphasize
that the drop in this area will lead to task being converted to a
bubble.
The existing indicator view is based on a single View and we modify the
bounds of the backgroud drawable.
If bubbles feature is enabled, use a FrameLayout for the View and add a
child view that is for the bubble bar indicator.
Animate bubble bar indicator together with the other indicator.

Bug: 388856523
Test: atest VisualIndicatorViewContainerTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: I892d10d2ce32271d3a502a301a7f342eb279085e

Former-commit-id: 0f4c171019b7365f993e90c1a532ed69ca86f779