Add an option to move floating bubbles to fullscreen
This option is in the bubble manage menu and triggers
the bubble to fullscreen transition.

Update the bubble to fullscreen transition to also check
for the BubbleExpandedView if BubbleBarExpandedView isn't
available.

Added a test for a bubble with phone expanded view.

Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Test: atest BubbleTransitionsTest
Test: manual - enable the flag, get a floating chat bubble,
               tap the manage menu, tap fullscreen
            => bubble moves to fullscreen
Bug: 400752049
Change-Id: I74dda034240eb2d83df6c324a1ff5e8862c49796

Former-commit-id: 0f82863dbccadad44d0c8f8413a34bce4236c953