Check system nav bar for custom IME Switcher vis
This enhances the onCustomImeSwitcherButtonRequestedVisible API to also
take into account the system navigation bar, in addition to the IME
navigation bar. The system bar is generally automatically shown whenever
the IME is shown, but this can be overriden through a config. Moreover,
the IME navigation bar can be disabled through a sysprop.

This takes into account both the config and the sysprop to provide a
more accurate response about the custom button requested visibility.

Flag: android.view.inputmethod.ime_switcher_revamp_api
Bug: 385018555
Test: atest
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_hideNavBarForKeyboard
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_imeNavigationBarDisabled
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_imeNavigationBarDisabled_hideNavBarForKeyboard
Change-Id: I1746e1f4cceb118140747c7ffb2b4266e3fee31e

Former-commit-id: 3a0f01da45e46545d81556c53bfacab4aba9c780