Fix flex split layout on landscape phones
Improves the calculation for apps that are extending offscreen.

Previously, offscreen bounds were calculated in a way that presumed symmetry (both left and right app having the same size in 50:50 split). But on phones in landscape orientation, we use a non-symmetric layout due to the camera cutout -- so the left app is actually larger in 50:50. This caused a bug.

Fixed in this CL by changing the way surface bounds are calculated.

Bug: 401396438
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: No longer fails test activityEmbeddingSplitSurfaceAreEven. Manually verified with smoke test and common flex split flows.
Change-Id: I1edbc2b052ab8f3dc770dcf789cc50536cc0a1d9

Former-commit-id: 124ab11a3770e62f188f55c11c32f1a1ea580d43