Default false for isSeamlessResizeEnabled
The PictureInPictureParams#isSeamlessResizeEnabled currently returns
true if an application does not set that explicitly.

We do see flicker from apps that are not seamless resize capable, while
not explicitly setting the flag to false. Doing the reverse, as in this
change, avoids most of those cases.

More context: go/default-seamless-resize-false

Flag: EXEMPT bugfix
Bug: 402783442
Test: atest PictureInPictureParamsTest
Change-Id: Ifcebd0d98b722a1c4b2adc5ecd0e18705eb6f9a1

Former-commit-id: 5f97146d824934397b72daf9cbe0bdcf0a2911dc