Desks: Do not re-create desks from other users
Fixes a crash when the re-create desk attempt fails for another user.
Prior to this change, the desk recreation was skipped as intented for
different users, but the DeskRecreationFactory returned the original
deskId, which caused the initializer to restore the desk with the old
deskId even if the desk was not re-created succesfully. Later on, when
trying to move a task into that non-existent desk, a crash would occur.

This change makes the factory return a null deskId in those cases, to
signal to the initializer not to restore the desk in the repository.

Proper multi-user desk restoration still needs to be implemented in the
future.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Fix: 402607975
Bug: 400984250
Test: in an HSUM build, drag a task into desktop - observe no crash
Change-Id: I9f7f6ed1485b665be7f4088859c884a6daa4b13c

Former-commit-id: b5dc18755c13a3ef0aa8e51d524199900a336a15