Add sign out option to user switcher
Add an action to the user switcher which allows the user to sign out
using the UserLogoutInteractor (either for the secondary user or for
HSUM devices to the system user). This option will only be displayed
if a logout is possible.

Test: atest UserSwitcherInteractorTest UserSwitcherViewModelTest KeyguardSecurityContainerTest StatusBarUserChipViewModelTest
Test: manual with flag enabled locally
Flag: com.android.systemui.user_switcher_add_sign_out_option
Bug: 381478261
Change-Id: Ie12be49353aa5cdc3eeb4db99cd7508bddbfcfbf

Former-commit-id: 1b75156d2fdcf21d605f30ab10fec7148237e021