Use `getWhen()` for AOD RONs
Meet's ongoing call notifs (which are *not* CallStyle, btw!) set a
`when` of 0, which is meant to default to the creation time of the
notification, so trigger that logic by calling `getWhen()` instead of
reading `when` when extracting the "when".

("when" is no longer a word. You're welcome.)

Fixes: 402954931
Test: atest PromotedNotificationContentExtractorImplTest
Test: manual, launch Meet call, go to AOD, verify chronometer
Flag: com.android.systemui.aod_ui_rich_ongoing
Change-Id: I30331aad61675029d69af79ab478df58eb4fff76

Former-commit-id: e98e04953e1588cf9ed137d395885f1f3c948ec5