Revert "Animate shade blur fallback"
This reverts commit 6ab6063cee2af18667dfbafc1e9025873cbd60ee [formerly e0677abea48704b418eec9aa7ea7d2c5f09518e3].

Reason for revert: Previously there was a bug so the blur isn't turned off when fallback happens. Now with blur turned off, the transparency doesn't get opaque until much later, causing everything to be seen through. With the animation, it makes things worse. Unless we find a better solution, let's revert this for now.

Bug: 388891313
Change-Id: I479c29cb95e4bb4cd182d996c4dcd06df051d851

Former-commit-id: 7cef087ce615ba976c88f0c65a4a84ef18e3ce0b