Fix NPE in adding/removing presentation
In automated testing, there seem to be some cases where a new
transition gets created for a presentation but immediately gets
aborted without moving to the collecting state (e.g. when sysui
gets killed or crashed when running a test, in which case no
transition player exists).

This change adds a check whether a display of interest is newly
added in the collecting transition in this operation.

Flag: com.android.window.flags.enable_presentation_for_connected_displays
Bug: 404107800
Test: PresentationTest
Change-Id: Icea03ae491c77b0c7e642ca822372d9999ed6916

Former-commit-id: 97ec86e906899e5af6d71dc9b2b530608ea48813