[flexiglass] Other dual shade overlay can show.
QuickSettingsShadeOverlayActionsViewModel and NotificationsShadeOverlayActionsViewModel both bind the swipe-down from an edge to a ShowOverlay with a hideCurrentOverlays of themselves as the overlay to hide. Unfortunately, this doesn't work because the hiding of the overlay causes a transition from that overlay to the Gone scene which triggers a visibility change to false with a log statement of "visible → invisible, reason: scene is Gone and no overlays are shown"; this cancels the transition to show the new overlay, resulting in this bug.

Changing ShowOverlay to ReplaceByOverlay seems to do the trick.

Fix: 394331336
Test: manually verified that dragging from the top edge of the other
overlay shows that overlay while collapsing the current one
Flag: com.android.systemui.scene_container

Change-Id: I46f4b6c0d3de5f6561ffbff685c2ee52c8574254

Former-commit-id: e4dd14c7efaa9e77c944583a0f2a58bfbccc18ab