Refine hub power button behavior
Transition directly from GONE -> HUB when applicable. This avoids the
intermediate transition to LOCKSCREEN which happens today, and removes
notification flickering. We also check for occlusion status to avoid
showing on top of a showWhenLocked activity.

We accomplish this by making the glanceable hub available always, not
only when locked. This ensures we can quickly switch to the communal
scene when the device is locked.

Fixes: 392969914
Fixes: 402645064
Fixes: 397447524
Flag: com.android.systemui.glanceable_hub_v2
Test: atest KeyguardTransitionScenariosTest
Test: atest CommunalInteractorTest
Test: atest FromGoneTransitionInteractorTest
Change-Id: I0c80c1b313133297a49de61df48a1f4a78616f81

Former-commit-id: 5485c74849efd86fd9e997fd647e4a7ad7dbd862