Add the "Connect a device" button on demand to the media item list.
This change adds a "Connect a device" item to the media item list upon
calling the getter method instead of as a response to the MediaRouter
callback. This allows to attach the item only when it's needed.

- `getMediaItemList()` returns the list with the "Connect a device"
button. This is used by the legacy mobile and TV UI.
- `getMediaItemList(false /* addConnectDeviceButton */)` returns the
list without the "Connect a device" button. This is used by the
refreshed mobile UI.

Bug: 386411078 Flag: EXEMPT refactor Test: atest
SystemUITests:com.android.systemui.media.dialog, atest
SystemUIGoogleScreenshotTests:com.android.systemui.media.dialog

Change-Id: Iac0df71eed25ee80501ab8e07a922edcb30401e1

Former-commit-id: c80df01652e9bb9cf88ce640f5314c36d1cf88a8