Make setCacheTestMode() a module API
Publish setCacheTestMode() as a module API so that mainline modules
can use it in their test code.  The API is flagged.

The new API forwards to the existing setTestMode() API.  It would be
preferable to reuse the existing setTestMode() API but this is
impossible under trunk-stable flagging rules.

Flag: android.os.ipc_data_cache_test_apis
Bug: 396173886
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
Change-Id: I4dac92526ef257a22d57e76b1b85f83ee0c71a7f

Former-commit-id: 06701b0e695c025fb839621969f0e631f899677d