Modifiying the rules that determine magnetic dismissibility
To avoid accidental dismissals, we also consider the magnitude and
direction of velocity to determine if a notification is dismissible
after it has been magnetically detached from its neighbors. When
detached, a notification snaps back when there is a fast enough fling in
the opposite direction relative to the detachment direction. In any
other case, the notification dismisses when let go.

Test: modified and added new Unit tests
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 405126597
Change-Id: Ic4c9074d2dc2e995c1b28b11529aa97395d1144e

Former-commit-id: b9b6d74e153859d710946df9f0ac2b3b2c001564